feat: create tabs for error views
