+++ arches_doorstep/src/arches_doorstep/etl_modules/ImportSingleCsvWithProcessing/views/Errors/ResourceErrorView.vue
@@ -1,0 +1,36 @@
+<script setup>
+
+</script>
+
+<template>
+    <div>
+        <Accordion>
+            <AccordionPanel>
+                <AccordionHeader>
+                    <div class="header-container">
+                        <div>Warnings</div>
+                        <div>0</div>
+                    </div>
+                </AccordionHeader>
+                <AccordionContent>
+                    Test
+                </AccordionContent>
+            </AccordionPanel>
+            <AccordionPanel>
+                <AccordionHeader>
+                    <div class="header-container">
+                        <div>Errors</div>
+                        <div>0</div>
+                    </div>
+                </AccordionHeader>
+                <AccordionContent>
+                    Test Errors
+                </AccordionContent>
+            </AccordionPanel>
+        </Accordion>
+    </div>
+</template>
+
+<style scoped>
+
+</style>
